Perform asset inlining on build, not on boot
