+++ layouts/partials/branding.ejs
@@ -3,7 +3,9 @@
     <% const shouldLink = locals.context !== 'home' %>
     <% if (shouldLink) { %><a id="branding-heading-link" href="/"><% } %>
       <span class="visually-hidden"><%- template('branding.org') %></span>
-      <img src=<%- locals.inlineAssets ? inlineAssets['branding.icon'] : template('branding.icon') %> class="logo" alt="<%- template('branding.org') %>">
+      <% if (!locals.useInlineStyles) { %>
+        <img src=<%- template('branding.icon') %> class="logo" alt="<%- template('branding.org') %>">
+      <% } %>
     <% if (shouldLink) { %></a><% } %>
   </h2>
   <h3 class="branding-label">
