+++ layouts/partials/header.ejs
@@ -2,8 +2,6 @@
   <div class="container">
     <%- include branding %>
     <%- include breadcrumb %>
-    <% if (!locals.inlineAssets) { %>
-      <%- include userTools %>
-    <% } %>
+    <%- include userTools %>
   </div>
 </header>
