+++ package.json
@@ -20,6 +20,7 @@
     "js-yaml": "^3.13.1",
     "lodash": "^4.17.11",
     "md5": "^2.2.1",
+    "mime-types": "^2.1.24",
     "moment": "^2.24.0",
     "node-sass": "^4.12.0",
     "passport": "^0.4.0",
@@ -50,7 +51,9 @@
   },
   "scripts": {
     "start": "node server/index.js",
-    "build": "node-sass --include-path custom/styles --include-path styles/partials --source-map public/css --source-map-contents --output public/css styles",
+    "build:main": "node-sass --include-path custom/styles --include-path styles/partials --source-map public/css/style.css.map --source-map-contents styles/style.scss public/css/style.css",
+    "build:inline": "./bin/build_inline_assets",
+    "build": "npm run build:main && npm run build:inline",
     "debug": "node -r dotenv/config --inspect --debug-brk server/index.js",
     "watch": "concurrently \"nodemon --inspect=0.0.0.0 -r dotenv/config -e js,ejs server/index.js\" \"nodemon -e scss --watch styles --watch custom/styles -x npm run build\"",
     "test": "NODE_ENV=test mocha test/**/*.test.js -r ./test/utils/bootstrap.js --recursive --exit",
